refactor(dao): Remove JobDAO and InMemoryDAO classes
